RT 59164: Query Config module for perl binary path
We need the perl binary used for this test run, which may not be the
same as the one found in 'env perl'.  By querying Config, we get
whatever options were passed to the configure script for perl.

Running the wordcount and analog tests are tricky because we must run
two external perl scripts, one of which needs to include the freshly
installed/tested Hadoop::Streaming libraries.

This is continued effort to fix build errors from the automated smoke
testers, who tend to run in minimal sandboxes, with multiple perl
versions installed.

https://rt.cpan.org/Ticket/Display.html?id=59164